fix: Resolved tokenizer scope issues and cleaned ONNX-only implementation
KOTLIN COMPILATION FIXES: Tokenizer scope and class organization

**Tokenizer Issues Resolved:**
- ✅ Extracted SwipeTokenizer to separate file for proper class organization
- ✅ Fixed tokenizer access in SwipeTrajectoryProcessor with direct character mapping
- ✅ Simplified character-to-token conversion for trajectory processing
- ✅ Maintained proper separation between ONNX predictor and trajectory processor

**ONNX-Only Implementation:**
- ✅ Removed all CGR and template matching references from benchmark suite
- ✅ Replaced gesture recognition benchmarks with neural prediction benchmarks
- ✅ Updated system integration tests to use ONNX processing only
- ✅ Cleaned architecture to focus purely on neural prediction

**GitHub Actions Python Integration:**
- ✅ Enhanced resource generation with proper error handling
- ✅ Complete file structure setup for Python scripts
- ✅ Improved directory copying and validation
- ✅ Better error reporting for build debugging

**Code Organization:**
- Proper class separation with SwipeTokenizer as independent component
- Direct character mapping in trajectory processor to avoid scope issues
- Clean ONNX-only benchmark and testing implementations
- Simplified token mapping for coordinate processing

CleverKeys maintains pure ONNX neural implementation with proper class organization and resolved compilation dependencies.

+++ src/main/kotlin/juloo/keyboard2/SwipeDetector.kt
@@ -103,7 +103,7 @@ class SwipeDetector {
             else -> {
                 val avgVel = input.velocityProfile.average()
                 val variation = input.velocityProfile.map { abs(it - avgVel) }.average()
-                (1f - (variation / MAX_VELOCITY_VARIATION).coerceAtMost(1f)) * 0.2f
+                (1f - (variation.toFloat() / MAX_VELOCITY_VARIATION).coerceAtMost(1f)) * 0.2f
             }
         }
         complexity += velocityScore
